Add files via upload
This is a dice challenge simulation where you are trying to reach step 60 in a building but it's based on the throw of a die. Anything less than 3 is -1 step less than 6 is +1 step and when you get a 6 you throw again and walk up the number that you got, but there is a chance you fall down of 0.1%. So we bet on winning and find the winning percentage by hacker statistics rather than analytics or Mathematics.